refactor(bot-node): extract helpers into telegrambot/lib
Three top-of-file helpers (formatErrorChain, parseStringArgList,
evalContextExpression) had been living at the head of bot-node.js purely
because they grew up there. They're pure (no closure capture from the
surrounding IIFE) and have nothing to do with Node-RED node wiring —
exactly the shape that belongs in telegrambot/lib alongside converter.js /
queue-manager.js / safe-stringify.js.

Two new lib files matching the existing one-file-per-concern style:
- telegrambot/lib/error-chain.js (formatErrorChain)
- telegrambot/lib/context-expression.js (evalContextExpression + its
  internal parseStringArgList helper)

bot-node.js now requires them at the top instead of defining them inline,
shedding ~149 lines including the module.exports.__test back-door that
existed solely so tests could reach the helpers. Tests now import from
the lib files directly, matching how test/lib/converter.test.js et al
already work.

The existing combined test file test/nodes/bot-node-helpers.test.js is
split along the same lines into test/lib/error-chain.test.js and
test/lib/context-expression.test.js. No test logic changes — pure
relocation. 239 passing.
