+++ telegrambot/nodes/bot-node.js
@@ -1,147 +1,5 @@
-// Walks an Error's .cause chain (Node 16+ standard) and any nested .errors arrays
-// (AggregateError) to collect the leaf error messages — i.e. the ones that actually
-// carry the useful diagnostic (typically the syscall-level message like
-// "connect ETIMEDOUT 149.154.166.110:443" rather than the intermediate library
-// wrappers' generic "AggregateError" / "RequestError" labels).
-//
-// The whole point is to get a one-line warn that tells an operator what to actually
-// fix (IPv4 vs IPv6, DNS, blocked port, ...) without needing to enable verbose
-// logging and read a 5-deep util.inspect dump.
-//
-// Returns a string formatted as: "<message1>; <message2>; ..." with consecutive
-// duplicates removed. Empty input -> empty string.
-function formatErrorChain(error) {
-    const seen = new Set();
-    const leaves = [];
-    function walk(e, depth) {
-        if (!e || typeof e !== 'object' || seen.has(e) || depth > 10) return;
-        seen.add(e);
-        const isAgg = Array.isArray(e.errors) && e.errors.length > 0;
-        const hasCause = e.cause && typeof e.cause === 'object';
-        if (isAgg) {
-            e.errors.forEach(function (inner) {
-                walk(inner, depth + 1);
-            });
-        }
-        if (hasCause) {
-            walk(e.cause, depth + 1);
-        }
-        if (!isAgg && !hasCause) {
-            // Prefer e.message; for plain string inputs, the string itself; else nothing.
-            // Avoid String(e) so a shape-less object doesn't show up as "[object Object]".
-            const msg = e.message || (typeof e === 'string' ? e : '');
-            if (msg) leaves.push(msg);
-        }
-    }
-    walk(error, 0);
-    // De-duplicate while preserving order.
-    const dedup = [];
-    leaves.forEach(function (m) {
-        if (dedup.indexOf(m) === -1) dedup.push(m);
-    });
-    if (dedup.length === 0) {
-        // Avoid the JS default "[object Object]" for shape-less inputs — fall back to
-        // the message if present, the raw string itself if the caller passed a string,
-        // else empty.
-        if (!error) return '';
-        if (typeof error === 'string') return error;
-        return error.message || '';
-    }
-    return dedup.join('; ');
-}
-
-// Parses a comma-separated list of single- or double-quoted string literals.
-// Returns the array of decoded strings, or null if the input is not a valid list of string literals.
-// Lifted to module scope so it can be unit-tested directly without a RED runtime.
-function parseStringArgList(input) {
-    const args = [];
-    let ok = true;
-    let i = 0;
-    const skipWs = function () {
-        while (i < input.length && /\s/.test(input[i])) i++;
-    };
-    skipWs();
-    while (ok && i < input.length) {
-        const quote = input[i];
-        if (quote !== '"' && quote !== "'") {
-            ok = false;
-        } else {
-            i++;
-            let val = '';
-            while (i < input.length && input[i] !== quote) {
-                if (input[i] === '\\' && i + 1 < input.length) {
-                    const next = input[i + 1];
-                    val += next === 'n' ? '\n' : next === 't' ? '\t' : next === 'r' ? '\r' : next;
-                    i += 2;
-                } else {
-                    val += input[i++];
-                }
-            }
-            if (input[i] !== quote) {
-                ok = false;
-            } else {
-                i++;
-                args.push(val);
-                skipWs();
-                if (i < input.length) {
-                    if (input[i] !== ',') {
-                        ok = false;
-                    } else {
-                        i++;
-                        skipWs();
-                    }
-                }
-            }
-        }
-    }
-    return ok ? args : null;
-}
-
-// Safely evaluates the small subset of expressions allowed in token / usernames / chatids fields.
-// Supported forms (see README):
-//   flow.get("key"[, "store"])     flow.keys()
-//   global.get("key"[, "store"])   global.keys()
-//   context.get("key"[, "store"])  context.keys()
-//   context.flow.get(...)          context.global.get(...)
-//   env.get("VAR")
-// Anything else evaluates to undefined.
-// Lifted to module scope so it can be unit-tested directly without a RED runtime.
-function evalContextExpression(node, expression) {
-    let result;
-    const trimmed = String(expression).trim();
-    const match = trimmed.match(/^(flow|global|context|env)(?:\.(flow|global))?\.(get|keys)\s*\(([\s\S]*)\)\s*$/);
-    if (match) {
-        const [, scope, subScope, method, argsRaw] = match;
-        const args = parseStringArgList(argsRaw);
-        if (args !== null) {
-            if (scope === 'env') {
-                if (!subScope && method === 'get' && args.length === 1) {
-                    try {
-                        result = node._flow.getSetting(args[0]);
-                    } catch (e) {
-                        // ignore — result stays undefined
-                    }
-                }
-            } else {
-                let target;
-                const ctx = node.context();
-                if (scope === 'context') {
-                    target = subScope ? ctx[subScope] : ctx;
-                } else if (!subScope) {
-                    target = ctx[scope];
-                }
-                if (target && typeof target[method] === 'function') {
-                    try {
-                        result = target[method](...args);
-                    } catch (e) {
-                        // ignore — result stays undefined
-                    }
-                }
-            }
-        }
-    }
-    return result;
-}
+const { formatErrorChain } = require('../lib/error-chain');
+const { evalContextExpression } = require('../lib/context-expression');
 
 module.exports = function (RED) {
     let telegramBot = require('node-telegram-bot-api');
@@ -1504,6 +1362,3 @@ module.exports = function (RED) {
 
     return TelegramBotNode;
 };
-
-// Exposed for unit tests. Not part of the public Node-RED API — do not consume from flows.
-module.exports.__test = { parseStringArgList, evalContextExpression, formatErrorChain };
